+++ packages/react-core/src/index.js
@@ -34,3 +34,5 @@ export { geojsonFeatures } from './filters/geojsonFeatures';
 
 export { GroupDateTypes } from './operations/GroupDateTypes';
 export { groupValuesByDateColumn } from './operations/groupByDate';
+
+export { DRAW_MODES, EDIT_MODES } from './utils/drawingToolConstants';