+++ packages/react-core/src/utils/drawingToolConstants.js
@@ -1,0 +1,10 @@
+export const DRAW_MODES = Object.freeze({
+  POLYGON: 'DrawPolygonMode',
+  RECTANGLE: 'DrawRectangleMode',
+  CIRCLE: 'DrawCircleFromCenterMode',
+  LASSO_TOOL: 'DrawPolygonByDraggingMode'
+});
+
+export const EDIT_MODES = Object.freeze({
+  EDIT: 'edit'
+});