+++ packages/react-widgets/src/index.js
@@ -6,6 +6,7 @@ export { default as PieWidget } from './widgets/PieWidget';
 export { default as LegendWidget } from './widgets/LegendWidget';
 export { default as ScatterPlotWidget } from './widgets/ScatterPlotWidget';
 export { default as TimeSeriesWidget } from './widgets/TimeSeriesWidget';
+export { default as DrawingToolWidget } from './widgets/DrawingToolWidget';
 export {
   getFormula,
   getHistogram,
@@ -14,3 +15,4 @@ export {
   getScatter
 } from './models';
 export { default as useSourceFilters } from './hooks/useSourceFilters';
+export { default as DrawingToolLayer } from './layers/DrawingToolLayer';